Move pin declaration to separate pins.rs
